Fix coordinate overflow for very large GPU volumes
This broke GPU volumes of more than 16GB, which we could start
allocating on GPUs with more than 48GB of memory.

Fixes #534.

+++ cuda/3d/cone_bp.cu
@@ -126,7 +126,7 @@ __global__ void dev_cone_BP(void* D_volData, unsigned int volPitch,
 		endZ = dims.iVolZ - startZ;
 
 	for(int i=0; i < endZ; i++)
-		volData[((startZ+i)*dims.iVolY+Y)*volPitch+X] += Z[i] * fOutputScale;
+		volData[(size_t)((startZ+i)*dims.iVolY+Y)*volPitch+X] += Z[i] * fOutputScale;
 } //End kernel
 
 
@@ -209,7 +209,7 @@ __global__ void dev_cone_BP_SS(void* D_volData, unsigned int volPitch, cudaTextu
 
 		}
 
-		volData[(Z*dims.iVolY+Y)*volPitch+X] += fVal * fOutputScale;
+		volData[(size_t)(Z*dims.iVolY+Y)*volPitch+X] += fVal * fOutputScale;
 	}
 }
